cpdetect: refactor Signatures
